refactor: split runner extension parity helpers
